backend_main/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Google Cloud Platform
google-cloud-firestore==2.13.1
google-cloud-storage==2.12.0
google-cloud-pubsub==2.18.4
google-cloud-bigquery==3.13.0
google-cloud-aiplatform==1.38.0

# Vertex AI
vertexai==1.38.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Data Processing
pandas==2.1.4
numpy==1.25.2

# PDF Processing
PyPDF2==3.0.1
python-multipart==0.0.6

# WebSocket for live data
websockets==12.0

# JSON Web Tokens
PyJWT==2.8.0

# Environment variables
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Date/Time utilities
python-dateutil==2.8.2

# Async utilities
asyncio-mqtt==0.16.1

# Testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx-test==0.1.0